requirements.txt
# Core LangChain/LangGraph Dependencies
langgraph==0.2.34
langchain==0.3.7
langchain-openai==0.2.8
langchain-core==0.3.15

# Bioinformatics Libraries
biopython==1.83

# Scientific Computing
numpy==1.24.3

# Visualization
matplotlib==3.8.2

# Testing
pytest==7.4.3

# Jupyter
jupyter==1.0.0
ipykernel==6.27.1

# Utilities
python-dotenv==1.0.0